[WIP] More refactoring (incomplete)
