+++ mypyc/ir/module_ir.py
@@ -5,7 +5,7 @@ from typing import List, Tuple, Dict
 from mypyc.common import JsonDict
 from mypyc.ir.ops import DeserMaps
 from mypyc.ir.rtypes import RType, deserialize_type
-from mypyc.ir.func_ir import FuncIR, FuncDecl, format_func
+from mypyc.ir.func_ir import FuncIR, FuncDecl
 from mypyc.ir.class_ir import ClassIR
 
 
@@ -82,12 +82,3 @@ def deserialize_modules(data: Dict[str, JsonDict], ctx: DeserMaps) -> Dict[str, 
 # ModulesIRs should also always be an *OrderedDict*, but if we
 # declared it that way we would need to put it in quotes everywhere...
 ModuleIRs = Dict[str, ModuleIR]
-
-
-def format_modules(modules: ModuleIRs) -> List[str]:
-    ops = []
-    for module in modules.values():
-        for fn in module.functions:
-            ops.extend(format_func(fn))
-            ops.append('')
-    return ops
